backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
pydantic==2.10.6
python-multipart==0.0.6
cryptography==46.0.3
groq==0.33.0
langchain==0.3.0
langchain-core==0.3.0
langgraph==0.2.20
langchain-openai==0.2.0
langchain-anthropic==0.2.0
langchain-groq==0.2.0
pydantic-settings==2.11.0
qdrant-client==1.12.2
mem0ai
ollama==0.6.0